+++ Source/Scene/ModelExperimental/ModelExperimental.js
@@ -6,7 +6,6 @@ import ClippingPlaneCollection from "../ClippingPlaneCollection.js";
 import defined from "../../Core/defined.js";
 import defer from "../../Core/defer.js";
 import defaultValue from "../../Core/defaultValue.js";
-import deprecationWarning from "../../Core/deprecationWarning.js";
 import DeveloperError from "../../Core/DeveloperError.js";
 import GltfLoader from "../GltfLoader.js";
 import ImageBasedLighting from "../ImageBasedLighting.js";
@@ -1477,8 +1476,7 @@ ModelExperimental.prototype.destroyResources = function () {
  * The model can be a traditional glTF asset with a .gltf extension or a Binary glTF using the .glb extension.
  *
  * @param {Object} options Object with the following properties:
- * @param {String|Resource|Uint8Array|Object} options.gltf A Resource/URL to a glTF/glb file, a binary glTF buffer, or a JSON object containing the glTF contents
- * @param {String|Resource} [options.url] The url to the .gltf file. Deprecated, use options.gltf instead.
+ * @param {String|Resource} options.url The url to the .gltf or .glb file.
  * @param {String|Resource} [options.basePath=''] The base path that paths in the glTF JSON are relative to.
  * @param {Matrix4} [options.modelMatrix=Matrix4.IDENTITY] The 4x4 transformation matrix that transforms the model from model to world coordinates.
  * @param {Number} [options.scale=1.0] A uniform scale applied to this model.
@@ -1516,19 +1514,15 @@ ModelExperimental.fromGltf = function (options) {
   options = defaultValue(options, defaultValue.EMPTY_OBJECT);
 
   //>>includeStart('debug', pragmas.debug);
-  if (!defined(options.gltf) && !defined(options.url)) {
-    throw new DeveloperError("options.gltf is required.");
+  if (!defined(options.url) && !defined(options.gltf)) {
+    throw new DeveloperError("options.url is required.");
   }
   //>>includeEnd('debug');
 
-  if (defined(options.url)) {
-    deprecationWarning(
-      "ModelExperimental.fromGltf",
-      "options.url is deprecated. Use options.gltf instead."
-    );
-  }
-
-  const gltf = defaultValue(options.gltf, options.url);
+  // options.gltf is used internally for 3D Tiles. It can be a Resource, a URL
+  // to a glTF/glb file, a binary glTF buffer, or a JSON object containing the
+  // glTF contents.
+  const gltf = defaultValue(options.url, options.gltf);
 
   const loaderOptions = {
     releaseGltfJson: options.releaseGltfJson,